+++ examples/eof_error.py
@@ -1,0 +1,3 @@
+value = input("Enter something: ")
+
+print(value)